week-5-and-6/requirements.txt
numpy>=1.24.0
sounddevice>=0.4.6
soundfile>=0.12.1
openai-whisper>=20231117
torch>=2.0.0
pyannote.audio>=2.1.1
groq>=0.4.0
fastapi>=0.104.0
pydantic>=2.0.0
pymongo>=4.5.0
bcrypt>=4.1.0
python-multipart>=0.0.6
uvicorn>=0.24.0
